% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbucket.R
\name{bitbucket_pwd}
\alias{bitbucket_pwd}
\title{Retrieve Bitbucket password}
\usage{
bitbucket_pwd(quiet = FALSE)
}
\description{
A bitbucket password
Looks in env var \code{BITBUCKET_PASSWORD}
}
\keyword{internal}
