% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbucket.R
\name{bitbucket_user}
\alias{bitbucket_user}
\title{Retrieve Bitbucket user.}
\usage{
bitbucket_user(quiet = FALSE)
}
\description{
A bitbucket user
Looks in env var \code{BITBUCKET_USERNAME}
}
\keyword{internal}
