% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable.R
\name{disable}
\alias{disable}
\title{Disable the use of Packrat in a Project}
\usage{
disable(project = NULL, restart = TRUE)
}
\arguments{
\item{project}{The directory in which packrat will be disabled (defaults to
the current working directory)}

\item{restart}{If \code{TRUE}, restart the R session after disabling packrat.}
}
\description{
Disable packrat within a project, reverting to the use of standard user
package libraries.
}
\note{
Disabling packrat for a project removes the packrat initialization code
from the .Rprofile file, resulting in the use of standard user package
libraries. Note that the \code{packrat} directory is not deleted, but remains
unused.

To re-enable the use of packrat for a project you can call the
\code{\link{init}} function.

The \code{restart} parameter will only result in a restart of R when the R
environment packrat is running within makes available a restart function via
\code{getOption("restart")}.
}
