% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{extlib}
\alias{extlib}
\title{Load an External Package}
\usage{
extlib(packages)
}
\description{
This function loads a non-packrat package that might be useful for development
while within a packrat project, but otherwise unused for a current project.
}
\details{
Unlike \code{\link{library}}, this function requires you to supply the package
name as a character vector, not a symbol.
}

