\name{flattenPackageRecords}
\alias{flattenPackageRecords}
\title{Returns a linear list of package records, sorted by name, with all
dependency information removed}
\usage{
  flattenPackageRecords(packageRecords)
}
\description{
  Returns a linear list of package records, sorted by name,
  with all dependency information removed
}

