% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{install_github}
\alias{install_github}
\title{Attempts to install a package directly from github.}
\usage{
install_github(repo, auth_user = NULL, auth_token = github_pat(), ...,
  dependencies = TRUE)
}
\arguments{
\item{repo}{Repository address in the format
\code{[username/]repo[/subdir][@ref|#pull]}.}

\item{auth_user}{Your account username if you're attempting to install
a package hosted in a private repository (and your username is different
to \code{username}). Currently disabled.}

\item{auth_token}{To install from a private repo, generate a personal
access token (PAT) in \url{https://github.com/settings/applications} and
supply to this argument. This is safer than using a password because
you can easily delete a PAT without affecting any others. Defaults to
the \code{GITHUB_PAT} environment variable. Currently disabled.}

\item{...}{Other arguments passed on to \code{install}.}

\item{dependencies}{By default, installs all dependencies so that you can
build vignettes and use all functionality of the package.}
}
\description{
This function is vectorised on \code{repo} so you can install multiple
packages in a single command.
}
\examples{
\dontrun{
install_github("roxygen")
install_github("wch/ggplot2")
install_github(c("rstudio/httpuv", "rstudio/shiny"))
install_github(c("devtools@devtools-1.4", "klutometis/roxygen#142", "mfrasca/r-logging/pkg))

# Update devtools to the latest version, on Linux and Mac
# On Windows, this won't work - see ?build_github_devtools
install_github("hadley/devtools")

# To install from a private repo, use auth_token with a token
# from https://github.com/settings/applications. You only need the
# repo scope. Best practice is to save your PAT in env var called
# GITHUB_PAT.
install_github("hadley/private", auth_token = "abc")

}
}

