% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{migrate}
\alias{migrate}
\title{Migrate to Packrat Mode}
\usage{
migrate(projDir = ".", ask = TRUE)
}
\arguments{
\item{projDir}{The project directory.}

\item{ask}{Boolean, ask before removing the \code{.Renviron} file?}
}
\description{
Migrate a project using the older version of packrat + packrat structure
to the newer version. This function will:
}
\details{
1. Move the \code{library} folder to the \code{packrat/lib} folder,
2. Move the \code{packrat.sources} folder to the \code{packrat/src} folder,
3. Move the \code{packrat.lock} file to \code{packrat/packrat.lock},
4. Remove the \code{.Renviron} file as it is no longer needed,
5. Update the package \code{.Rprofile}.
}

