% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{packrat-external}
\alias{extlib}
\alias{packrat-external}
\alias{with_extlib}
\title{Managing External Libraries}
\usage{
with_extlib(packages, expr, envir = parent.frame())

extlib(packages)
}
\arguments{
\item{packages}{A set of library names (as a character vector) to load for
the duration of evaluation of \code{expr}.}

\item{expr}{An \R expression.}

\item{envir}{An environment in which the expression is evaluated.}
}
\description{
These functions provide a mechanism for (temporarily) using packages outside of the
packrat private library.
}
\examples{
\dontrun{
with_extlib("lattice", xyplot(1 ~ 1))
}
}

