% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{packratOn}
\alias{packratMode}
\alias{packratOff}
\alias{packratOn}
\title{Packrat Mode}
\usage{
packratOn(projDir = ".", autoSnapshot = TRUE, bootstrap = FALSE)

packratOff(projDir = NULL)

packratMode(projDir = NULL, autoSnapshot = TRUE, bootstrap = FALSE)
}
\arguments{
\item{projDir}{The directory in which packrat mode is launched -- this is
where local libraries will be used and updated.}

\item{autoSnapshot}{Whether or not we should use automatic snapshotting.}

\item{bootstrap}{Whether or not we should try to bootstrap a project directory
that has not yet been packified.}
}
\description{
Use these functions to switch \code{packrat} mode on and off. When within
\code{packrat} mode, the library is set to use a local packrat library
within the current project.
}

