\name{pkgNamesAndVersions}
\alias{pkgNamesAndVersions}
\title{Filters out all record properties except name and version. Dependencies
are dropped.}
\usage{
  pkgNamesAndVersions(packageRecords)
}
\description{
  Filters out all record properties except name and
  version. Dependencies are dropped.
}

