% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{remove}
\alias{remove}
\title{Remove Packrat from a Project}
\usage{
remove(project = NULL, restart = TRUE)
}
\arguments{
\item{project}{The directory which packrat will be removed from (defaults
to the current working directory)}

\item{restart}{If \code{TRUE}, restart the R session after restoring.}
}
\description{
Remove packrat from the specified project, reverting to the use of standard
user package libraries.
}
\note{
The \code{restart} parmaeter will only result in a restart of R when the
R environment packrat is running within makes available a restart function
via \code{getOption("restart")}.
}

