% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranlike-repositories.R
\name{repos_create}
\alias{repos_create}
\title{Create a Local, CRAN-like Repository}
\usage{
repos_create(path, name = basename(path), add = TRUE)
}
\arguments{
\item{path}{Path to a local CRAN-like repository.}

\item{name}{The name to assign to the repository. Defaults to the
directory name in which the reopsitory is created.}

\item{add}{Add this new repository to the current set of repositories?}
}
\description{
Generate a local CRAN-like repository which can be
used to store and distribute \R packages.
}
