% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranlike-repositories.R
\name{repos_upload}
\alias{repos_upload}
\title{Upload a Package to a Local CRAN-like Repository}
\usage{
repos_upload(package, to, ...)
}
\arguments{
\item{package}{Path to a package tarball. The tarball should be
created by \code{R CMD build}; alternatively, it can be the path
to a folder containing the source code for a package (which
will then be built with \code{R CMD build}) and then uploaded
to the local repository.}

\item{to}{The name of the CRAN-like repository. It (currently) must
be a local (on-disk) CRAN repository.}

\item{...}{Optional arguments passed to \code{R CMD build}.}
}
\description{
Upload a Package to a Local CRAN-like Repository
}
