% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cranlike-repositories.R
\name{repository-management}
\alias{add_repos}
\alias{repository-management}
\alias{set_repos}
\title{Add a Repository}
\usage{
add_repos(..., overwrite = FALSE)

set_repos(...)
}
\arguments{
\item{...}{Named arguments of the form \code{<repoName> = <pathToRepo>}}

\item{overwrite}{Boolean; overwrite if a repository with the given name
  already exists?}
}
\description{
Add a repository to the set of currently available repositories. This is
effectively an easier-to-use wrapper over interacting with the
\code{"repos"} option, which is otherwise set with \code{options(repos = ...)}.
}

