% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{srcDir}
\alias{srcDir}
\title{Show the path to the current private sources}
\usage{
srcDir(projDir = ".")
}
\arguments{
\item{projDir}{The project directory. Defaults to current working
directory.}
}
\value{
A character vector containing the path to the private package sources. The path
is not guaranteed to exist on disk.
}
\description{
Returns the path to the private package sources used by packrat.
}
\note{
The private package library is normally created by \code{\link{bootstrap}}.
}
\examples{
# Show the library directory for the current working directory
srcDir()
}

