% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{unbundle}
\alias{unbundle}
\title{Unbundle a Packrat Project}
\usage{
unbundle(bundle, where, ...)
}
\arguments{
\item{bundle}{Path to the bundled file.}

\item{where}{The directory where we will unbundle the project.}

\item{...}{Optional arguments passed to \code{\link{tar}}.}
}
\description{
Unbundle a previously \code{\link{bundle}}d project.
}

