% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cranlike-repositories.R
\name{upload_package}
\alias{upload_package}
\title{Upload a Package to a Local CRAN-like Repository}
\usage{
upload_package(package, to)
}
\arguments{
\item{package}{Path to a package tarball. The tarball should be
created by \code{R CMD build}; alternatively, it can be the path
to a folder containing the source code for a package (which
will then be built with \code{R CMD build}) and then uploaded
to the local repository.}
}
\description{
Upload a Package to a Local CRAN-like Repository
}

