% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{withExtLibs}
\alias{withExtLibs}
\title{Temporarily load External Libraries for Evaluation of an Expression}
\usage{
withExtLibs(libs, expr)
}
\arguments{
\item{libs}{A set of library names (as a character vector) to load for
the duration of evaluation of \code{expr}.}

\item{expr}{An \R expression.}
}
\description{
This can be useful if you need to call a function from a library
not managed by \code{packrat}, e.g. \code{devtools}.
}

