% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{clean}
\alias{clean}
\title{Remove unused packages}
\usage{
clean(projDir = NULL, lib.loc = libDir(projDir), prompt = interactive())
}
\arguments{
\item{projDir}{The project directory. Defaults to current working
directory.}

\item{lib.loc}{The library to clean. Defaults to the private package library
associated with the project directory.}

\item{prompt}{\code{TRUE} to prompt before removing packages, \code{FALSE} to
remove packages immediately.}
}
\description{
Remove unused packages from the given library.
}
\details{
Detects and removes orphaned packages. Orphaned packages are those that meet
the following criteria:
\itemize{
  \item Installed in the library
  \item Not directly used by any \R code in the project
  \item Not a dependency of any non-orphaned package
}
If \code{clean} wants to remove a package but your project actually needs the
package, add a statement such as \code{\link{require}(package-name)} to any .R
file in your project's directory.
}
\seealso{
\code{\link{appDependencies}} for an explanation of how dependencies are detected.
}

