% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cranlike-repositories.R
\name{create_repo}
\alias{create_repo}
\title{Create a Local, CRAN-like Repository}
\usage{
create_repo(path, name = basename(path), add = TRUE)
}
\arguments{
\item{path}{Path to the local CRAN-like repository.}

\item{name}{The name to assign to the repository. Defaults to the
directory name in which the reopsitory is created.}

\item{add}{Add this new repository to the current set of repositories?}
}
\description{
This function generates a local CRAN-like repository which can be
used to store and distribute \R packages.
}

