% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{extlib}
\alias{extlib}
\title{Load an External Package}
\usage{
extlib(packages)
}
\arguments{
\item{packages}{A list of \R packages to load from the default user library.}
}
\description{
Load a package from the user (non-private) library.
}
\details{
Unlike \code{\link{library}}, this function requires you to supply the package
name as a character vector, not a symbol.
}

