% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{install_github}
\alias{install_github}
\title{Attempts to install a package directly from github.}
\usage{
install_github(repo, ..., dependencies = TRUE)
}
\arguments{
\item{repo}{Repository address in the format
\code{[username/]repo[/subdir][@ref|#pull]}.}

\item{...}{Other arguments passed on to \code{install}.}

\item{dependencies}{By default, installs all dependencies so that you can
build vignettes and use all functionality of the package.}
}
\description{
This function is a bare-bones version of \code{devtools::install_github}.
Tools for authentication are not available, but the main interface through
the \code{repo} argument is preserved.
}
\examples{
\dontrun{
install_github("roxygen")
install_github("wch/ggplot2")
install_github(c("rstudio/httpuv", "rstudio/shiny"))
install_github(c("devtools@devtools-1.4", "klutometis/roxygen#142", "mfrasca/r-logging/pkg))

# Update devtools to the latest version, on Linux and Mac
# On Windows, this won't work - see ?build_github_devtools
install_github("hadley/devtools")

# To install from a private repo, use auth_token with a token
# from https://github.com/settings/applications. You only need the
# repo scope. Best practice is to save your PAT in env var called
# GITHUB_PAT.
install_github("hadley/private", auth_token = "abc")

}
}
\seealso{
Other package.installation: \code{\link{install_url}}
}

