% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/migrate-packrat.R
\name{migrate}
\alias{migrate}
\title{Migrate to Packrat Mode}
\usage{
migrate(project = ".", ask = TRUE)
}
\arguments{
\item{project}{The project directory.}

\item{ask}{Boolean, ask before removing the \code{.Renviron} file?}
}
\description{
Migrate a project using the older version of packrat + packrat structure
to the newer version.
}
\details{
This function will:

\enumerate{
\item Move the \code{library} folder to the \code{packrat/lib} folder,
\item Move the \code{packrat.sources} folder to the \code{packrat/src} folder,
\item Move the \code{packrat.lock} file to \code{packrat/packrat.lock},
\item Remove the \code{.Renviron} file as it is no longer needed,
\item Update the package \code{.Rprofile}.
}
}

