% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packrat.R
\name{packify}
\alias{packify}
\title{Automatically Enter Packrat Mode on Startup}
\usage{
packify(project = NULL, quiet = FALSE)
}
\arguments{
\item{project}{The directory in which to install the \code{.Rprofile} file.}

\item{quiet}{Be chatty?}
}
\description{
Install/augment the \code{.Rprofile} in a project, so that all \R sessions
started in this directory enter \code{packrat mode}, and use the local
project library.
}
\details{
It is not normally necessary to call \code{packify} directly; these files are
normally installed by \code{\link{init}}. \code{packify} can be used to
restore the files if they are missing (for instance, if they were not added to
source control, or were accidentally removed).

You'll need to restart \R in the specified directory after running
\code{packify} in order to start using the private package library.
}
