% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R, R/paths.R
\name{packrat-external}
\alias{packrat-external}
\alias{with_extlib}
\alias{packrat-external}
\alias{extlib}
\alias{packrat-external}
\alias{user_lib}
\alias{packrat-external}
\alias{packrat_lib}
\title{Managing External Libraries}
\usage{
with_extlib(packages, expr, envir = parent.frame())

extlib(packages)

user_lib()

packrat_lib()
}
\arguments{
\item{packages}{A set of package names (as a character vector) to load for
the duration of evaluation of \code{expr}.}

\item{expr}{An \R expression.}

\item{envir}{An environment in which the expression is evaluated.}
}
\description{
These functions provide a mechanism for (temporarily) using packages outside
of the packrat private library. The packages are searched within the 'default'
libraries; that is, the libraries that would be available upon launching a new
\R session.
}
\examples{
\dontrun{
with_extlib("lattice", xyplot(1 ~ 1))
}
}
