% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{packrat-mode}
\alias{packrat-mode}
\alias{packrat_mode}
\alias{packrat_off}
\alias{packrat_on}
\title{Packrat Mode}
\usage{
packrat_on(projDir = ".", auto.snapshot = TRUE, bootstrap = FALSE)

packrat_off(projDir = NULL)

packrat_mode(projDir = NULL, auto.snapshot = TRUE, bootstrap = FALSE)
}
\arguments{
\item{projDir}{The directory in which packrat mode is launched -- this is
where local libraries will be used and updated.}

\item{auto.snapshot}{Whether or not we should use automatic snapshotting.}

\item{bootstrap}{Whether or not we should try to bootstrap a project directory
that has not yet been packified.}
}
\description{
Use these functions to switch \code{packrat} mode on and off. When within
\code{packrat} mode, the \R session will use the private library generated
for the current project.
}
\details{
\code{packrat_mode} is used to toggle packrat mode on and off, while
\code{packrat_on} and \code{packrat_on} can be used to force packrat mode
on and off, respectively.
}

