% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{packrat-mode}
\alias{packrat-mode}
\alias{packrat_mode}
\title{Packrat Mode}
\usage{
packrat_mode(on = NULL, projDir = NULL, bootstrap = FALSE,
  auto.snapshot = TRUE, clean.search.path = TRUE)
}
\arguments{
\item{on}{Turn packrat mode on (\code{TRUE}) or off (\code{FALSE}). If omitted, packrat mode
will be toggled.}

\item{projDir}{The directory in which packrat mode is launched -- this is
where local libraries will be used and updated.}

\item{auto.snapshot}{Whether or not we should use automatic snapshotting.}

\item{bootstrap}{Whether or not we should try to bootstrap a project directory
that has not yet been packified.}

\item{clean.search.path}{Detach and unload any packages loaded from a user library before
entering packrat mode?}
}
\description{
Use these functions to switch \code{packrat} mode on and off. When within
\code{packrat} mode, the \R session will use the private library generated
for the current project.
}
\details{
\code{packrat_mode} is used to toggle packrat mode on and off, while
\code{packrat_on} and \code{packrat_on} can be used to force packrat mode
on and off, respectively.
}

