% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{packrat-options}
\alias{get_opts}
\alias{packrat-options}
\alias{set_opts}
\title{Get/set packrat project options}
\usage{
get_opts(options = NULL)

set_opts(...)
}
\arguments{
\item{options}{A character vector of valid option names.}

\item{...}{Entries of the form \code{key = value}, used for setting packrat project options.}
}
\description{
Get and set a number of options for the current packrat-managed project.
}
\details{
\itemize{
\item \code{auto.snapshot}: Perform automatic, asynchronous snapshots when running interactively?
  (\code{TRUE} / \code{FALSE})
\item \code{git.ignore.lib}: Add the packrat private library to the git ignore? (\code{TRUE} / \code{FALSE})
\item \code{git.ignore.src}: Add the packrat private sources to the git ignore? (\code{TRUE} / \code{FALSE})
\item \code{svn.ignore.lib}: Add the packrat private library to the SVN ignore? (\code{TRUE} / \code{FALSE})
\item \code{svn.ignore.src}: Add the packrat private sources to the SVN ignore? (\code{TRUE} / \code{FALSE})
}
}

