% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{packrat-options}
\alias{get_opts}
\alias{packrat-options}
\alias{set_opts}
\title{Get/set packrat project options}
\usage{
get_opts(options = NULL, simplify = TRUE, project = NULL)

set_opts(..., project = NULL)
}
\arguments{
\item{options}{A character vector of valid option names.}

\item{simplify}{Boolean; \code{unlist} the returned options? Useful for when retrieving
a single option.}

\item{project}{The project directory. When in packrat mode, defaults to the current project;
otherwise, defaults to the current working directory.}

\item{...}{Entries of the form \code{key = value}, used for setting packrat project options.}
}
\description{
Get and set options for the current packrat-managed project.
}
\details{
\itemize{
\item \code{auto.snapshot}: Perform automatic, asynchronous snapshots when running interactively?
  (\code{TRUE} / \code{FALSE})
\item \code{vcs.ignore.lib}:
  Add the packrat private library to your version control system ignore? (\code{TRUE} / \code{FALSE})
\item \code{vcs.ignore.src}:
  Add the packrat private sources to your version control system ignore? (\code{TRUE} / \code{FALSE})
\item \code{print.banner.on.startup}:
  Print the banner on startup? Can be one of \code{TRUE} (always print),
  \code{FALSE} (never print), and \code{'console'} (only print if session
  started from the console).
}
}

