% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{packrat_mode}
\alias{packrat_mode}
\title{Packrat Mode}
\usage{
packrat_mode(projDir = ".")
}
\arguments{
\item{projDir}{The directory in which packrat mode is launched -- this is
where local libraries will be used and updated.}
}
\description{
Use this function to switch \code{packrat} mode on and off. When within
\code{packrat} mode, the library is set to use a local packrat library
within \code{file.path(projDir, ".packrat/lib")}.
}

