% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranlike-repositories.R
\name{repository-management}
\alias{repository-management}
\alias{repos_add}
\alias{repos_add_local}
\alias{repos_set}
\alias{repos_set_local}
\alias{repos_remove}
\alias{repos_list}
\title{Add a Repository}
\usage{
repos_add(..., overwrite = FALSE)

repos_add_local(..., overwrite = FALSE)

repos_set(...)

repos_set_local(...)

repos_remove(names)

repos_list()
}
\arguments{
\item{...}{Named arguments of the form \code{<repoName> = <pathToRepo>}.}

\item{overwrite}{Boolean; overwrite if a repository with the given name
already exists?}

\item{names}{The names of repositories (as exist in e.g.
\code{names(getOption("repos"))}).}
}
\description{
Add a repository to the set of currently available repositories. This is
effectively an easier-to-use wrapper over interacting with the
\code{"repos"} option, which is otherwise set with \code{options(repos = ...)}.
}
\details{
\code{repos_add_local} is used for adding file-based repositories; that is,
CRAN repositories that live locally on disk and not on the internet / local network.
}
