% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{unbundle}
\alias{unbundle}
\title{Unbundle a Packrat Project}
\usage{
unbundle(bundle, where, ..., restore = TRUE)
}
\arguments{
\item{bundle}{Path to the bundled file.}

\item{where}{The directory where we will unbundle the project.}

\item{...}{Optional arguments passed to \code{\link{tar}}.}

\item{restore}{Boolean; should we \code{\link{restore}} the library
after \code{unbundle}-ing the project?}
}
\description{
Unbundle a previously \code{\link{bundle}}d project.
}
