% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packrat.R
\name{unused_packages}
\alias{unused_packages}
\title{Find Unused Packages in a Project}
\usage{
unused_packages(project = NULL, lib.loc = libDir(project))
}
\arguments{
\item{project}{The project directory.}

\item{lib.loc}{The library to check.}
}
\description{
Unused packages are those still contained within your project library, but
are unused in your project.
}
